Pretty much implementing the final AI things for fnaf 2
The Animatronics can now move around and be seen on the cameras, they dont jumpscare you yet and they dont appear in locations such as your vents, hallway or in the office itself. The only fully operational animatronic is the puppet which will actually jumpscare you. its important to note though that the jumpscares still need some work.